refactor: 서바이벌 스터디 경로 연결
